create separate book map file
